refactor: gameManager 패키지 변경
