+++ src/main/kotlin/omok/Application.kt
@@ -1,7 +1,6 @@
 package omok
 
 import omok.controller.Controller
-import omok.model.GameManager
 
 fun main() {
     val controller = Controller(GameManager())
