+++ src/main/kotlin/omok/controller/Controller.kt
@@ -1,7 +1,7 @@
 package omok.controller
 
+import omok.GameManager
 import omok.model.Board
-import omok.model.GameManager
 import omok.model.GameState
 import omok.model.Position
 import omok.view.BoardView
